+++ src/ui/manualrenderwidget.cpp
@@ -37,6 +37,17 @@ ManualRenderWidget::ManualRenderWidget(QWidget* parent) : QWidget(parent), ui(ne
                 }
             });
 
+    connect(ui->browseChaptersManualButton, &QPushButton::clicked, this,
+            [this]()
+            {
+                QString path = QFileDialog::getOpenFileName(this, "Файл глав Matroska XML", "",
+                                                            "XML (*.xml);;Все файлы (*)");
+                if (!path.isEmpty())
+                {
+                    ui->chaptersExternalPathManualEdit->setText(path);
+                }
+            });
+
     updateHardsubOptions();
 }
 
@@ -210,6 +221,8 @@ QVariantMap ManualRenderWidget::getParameters() const
         useConcatTb = concatTbCheckBox->isChecked();
     }
     params["useConcatTb"] = useConcatTb;
+    params["chaptersExternalPath"] = ui->chaptersExternalPathManualEdit->text().trimmed();
+    params["transferEmbeddedChapters"] = ui->transferEmbeddedChaptersCheckBox->isChecked();
 
     if (useHardsub)
     {